YAY-22 refactor: remove ProfileView widget and move ProfileHeader to widget layer
