+++ src/entities/user/index.ts
@@ -1,5 +1,4 @@
 export { UserStats } from './ui/UserStats'
-export { ProfileHeader } from './ui/ProfileHeader/ProfileHeader'
 export * from './model/user.types'
 export * from './api/user.service'
 export * from './model/userEndpoints'
