+++ src/widgets/Profile/ui/ProfileHeader/ProfileHeader.tsx
@@ -1,9 +1,8 @@
-import { UserStats } from '../UserStats'
 import { Typography } from '@/shared/components/Typography'
 import { PostOutlineIcon } from '@/shared/icons'
 import { Button } from '@/shared/components/Button'
 import Link from 'next/link'
-import { useGetUserByLoginQuery } from '@/entities/user'
+import { useGetUserByLoginQuery, UserStats } from '@/entities/user'
 import React from 'react'
 import { useMeQuery } from '@/features/auth/api'
 import { Avatar } from '@/shared/components/Avatar'
